// src/api/constants.ts
export const enum TaskCategory {
  StartOnboarding = 'start_onboarding',
  FinishOnboarding = 'finish_onboarding',
  ShareVibe = 'share_vibe',
  AdView = 'ad_view',
  PublishPhotos = 'publish_photos',
  PublishVibelists = 'publish_vibelists',
  BringFriends = 'bring_friends',
  SubscribeToUsers = 'subscribe_to_users',
  Likes = 'likes',
  Comments = 'comments',
  Subscribers = 'subscribers',
  LoginStreak = 'login_streak',
  SubscribeChannel = 'subscribe_channel',
  SubscribeChat = 'subscribe_chat',
  SubscribeInstagram = 'subscribe_instagram',
  SubscribeYoutube = 'subscribe_youtube',
  SubscribeTikTok = 'subscribe_tiktok',
  SubscribeX = 'subscribe_x',
}

export const enum TaskCategorySocialTg {
  SubscribeChannel = 'subscribe_channel',
  SubscribeChat = 'subscribe_chat',
}

export const enum TaskCategorySocialExt {
  SubscribeInstagram = 'subscribe_instagram',
  SubscribeYoutube = 'subscribe_youtube',
  SubscribeTikTok = 'subscribe_tiktok',
  SubscribeX = 'subscribe_x',
}

export const AD_INTERVAL = 45;